5.10 Попытка исправить замечания
