+++ src/components/commons/list-films/list-films.connect.js
@@ -1,0 +1,4 @@
+export const mapStateToProps = (state) => ({
+  films: state.films,
+  activeGenre: state.activeGenre
+});
