+++ src/components/commons/list-genres/list-genres.jsx
@@ -1,7 +1,7 @@
 import {connect} from "react-redux";
-import {ActionCreator} from "../../../store/action";
+import {mapStateToProps, mapDispatchToProps} from "./list-genres.connect";
 
-const ListGenres = ({genres, activeGenre, changeActiveFilter, changeFilmList}) => {
+const ListGenres = ({genres, activeGenre, changeActiveFilter}) => {
   return (
     <ul className="catalog__genres-list">
       {genres.map((genre, i) =>
@@ -11,7 +11,6 @@ const ListGenres = ({genres, activeGenre, changeActiveFilter, changeFilmList}) =
           onClick={(evt) => {
             evt.preventDefault();
             changeActiveFilter(genre);
-            changeFilmList(genre);
           }}
         >
           <a href="#" className="catalog__genres-link">{genre}</a>
@@ -21,25 +20,10 @@ const ListGenres = ({genres, activeGenre, changeActiveFilter, changeFilmList}) =
   );
 };
 
-const mapStateToProps = (state) => ({
-  genres: state.genres,
-  activeGenre: state.activeGenre
-});
-
-const mapDispatchToProps = (dispatch) => ({
-  changeActiveFilter(genreForFilter) {
-    dispatch(ActionCreator.changeActiveFilter(genreForFilter));
-  },
-  changeFilmList(genreForFilter) {
-    dispatch(ActionCreator.changeFilmList(genreForFilter));
-  }
-});
-
 ListGenres.propTypes = {
-  genres: PropTypes.arrayOf(PropTypes.string.isRequired),
+  genres: PropTypes.arrayOf(PropTypes.string.isRequired).isRequired,
   activeGenre: PropTypes.string.isRequired,
-  changeActiveFilter: PropTypes.func.isRequired,
-  changeFilmList: PropTypes.func.isRequired
+  changeActiveFilter: PropTypes.func.isRequired
 };
 
 export {ListGenres};
