+++ src/components/screens/film-screen/film-screen.connect.js
@@ -1,0 +1,3 @@
+export const mapStateToProps = (state) => ({
+  films: state.films
+});
