+++ src/store/reducer.js
@@ -1,13 +1,20 @@
 import {extend} from "../utils/common";
 import {ActionType} from "./action";
-import {getFilmsByGenre, genres} from "../utils/films";
 import {allFilms} from "../mocks/films";
 import {FilmGenres} from "../consts";
+
+const getUniqueGenresArray = (films) => {
+  const dirtyFilters = [];
+  films.forEach((film) => dirtyFilters.push((film.genres)));
+  return [FilmGenres.ALL_GENRES, ...new Set(dirtyFilters.flat())];
+};
+
+const genres = getUniqueGenresArray(allFilms);
 
 const initialState = {
   genres,
   activeGenre: FilmGenres.ALL_GENRES,
-  filteredFilms: allFilms
+  films: allFilms
 };
 
 const reducer = (state = initialState, action) => {
@@ -15,10 +22,6 @@ const reducer = (state = initialState, action) => {
     case ActionType.CHANGE_FILTER_BY_GENRE:
       return extend(state, {
         activeGenre: action.payload
-      });
-    case ActionType.GET_FILMS_BY_GENRE:
-      return extend(state, {
-        filteredFilms: getFilmsByGenre(allFilms, action.payload)
       });
     default:
       return state;
